1.1 changelog (#1581)
Ended up pivoting to a simple changelog approach for the 1.1 release
notes. This is more sustainable and motivates us to write up robust docs
that we can link out to for new features and changes.

Post actual release, we can link out to the changelog on github w/ nitty
gritty details.

Also fixing some broken links :)

---------

Co-authored-by: Christian Bromann <git@bromann.dev>
Co-authored-by: Chester Curme <chester.curme@gmail.com>
Co-authored-by: Mason Daugherty <mason@langchain.dev>